Rename gcr -> gcrane
